Update component paths; Default main styling; Add username to login sessionStorage response;
