+++ application/frontend/src/components/Search.jsx
@@ -4,13 +4,14 @@ import useUserProfileStore from "../stores/userProfileStore";
 import axios from "axios";
 import { HOST_PATH } from "../scripts/constants";
 
-
 const Search = () => {
   const [inputValue, setInputValue] = useState("");
   const [filteredUsernames, setFilteredUsernames] = useState([]);
   const [error, setError] = useState(null);
   const { setProfileId } = useUserProfileStore();
   const navigate = useNavigate();
+
+  console.log(HOST_PATH);
 
   // Fetch usernames from the backend as user types
   const fetchUsernames = async (searchTerm) => {
@@ -68,10 +69,8 @@ const Search = () => {
       {inputValue && filteredUsernames.length > 0 && (
         <ul className="username-search-results">
           {filteredUsernames.map((user, index) => (
-            <li key={index}>
-              <button onClick={() => handleUserClick(user.id)}>
-                {user.username}
-              </button>
+            <li key={index} onClick={() => handleUserClick(user.id)}>
+              {user.username}
             </li>
           ))}
         </ul>
