+++ application/frontend/src/components/pages/Home.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-import { HOST_PATH } from "../scripts/constants";
+import { HOST_PATH } from "../../scripts/constants";
 
 export default function Home() {
   const [lobbyGames, setLobbyGames] = useState([]);
