+++ application/frontend/src/components/pages/Lobby.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-import { HOST_PATH } from "../scripts/constants";
+import { HOST_PATH } from "../../scripts/constants";
 
 export default function Lobby() {
   const [lobbyGames, setLobbyGames] = useState([]);
@@ -23,7 +23,7 @@ export default function Lobby() {
   }, []);
 
   return (
-    <main className="main-lobby">
+    <main className="main-default main-lobby">
       <article className="lobby-article">
         <h2>Lobby</h2>
         {lobbyGames.length > 0 ? (
