+++ application/frontend/src/scripts/constants.js
@@ -1,1 +1,1 @@
-export const HOST_PATH = "localhost:8000/api"
+export const HOST_PATH = "http://localhost:8000/api"
