+++ application/frontend/src/styles/_index.scss
@@ -11,6 +11,7 @@
 @forward "./animations/animations.scss";
 
 // Pages
+@forward "./pages/pages.scss";
 @forward "./pages/dashboard.scss";
 @forward "./pages/profile.scss";
 @forward "./pages/home.scss";
