+++ application/frontend/src/styles/pages/home.scss
@@ -22,8 +22,8 @@
 
 .preview-placeholder {
     width: 100%;
-    min-height: calc(80vh - constants.$navigation-height);
-    max-height: calc(80vh - constants.$navigation-height);
+    min-height: calc(90vh - constants.$navigation-height);
+    max-height: calc(90vh - constants.$navigation-height);
     background-color: #e0e0e0;
     border-radius: 8px;
     margin: 20px 0;
