+++ application/frontend/src/styles/pages/pages.scss
@@ -1,0 +1,5 @@
+@use "../constants/constants";
+
+.main-default {
+    margin: 0 constants.$default-margin
+}
